+++ src/main/java/com/example/voebb/model/dto/user/UserRegistrationDTO.java
@@ -1,4 +1,4 @@
-package com.example.voebb.model.dto;
+package com.example.voebb.model.dto.user;
 
 //TODO: add validation in this record
 public record UserRegistrationDTO(
