+++ src/main/resources/templates/admin/fragments/add-product-modal.html
@@ -8,7 +8,7 @@
                     <button type="button" class="btn-close" data-bs-dismiss="modal"></button>
                 </div>
 
-                <form th:action="@{/admin/products}" method="post">
+                <form th:action="@{/templates/admin/products}" method="post">
                     <div class="modal-body">
                         <div class="row g-3">
