+++ src/main/resources/templates/user/fragments/product-card.html
